# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ChapterBrief": {
        "description": "章节列表项：**不含正文**（坑 11，TC-13）。",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "done"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "word_count": {
            "default": 0,
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "seq",
          "updated_at"
        ],
        "title": "ChapterBrief",
        "type": "object"
      },
      "ChapterCreate": {
        "properties": {
          "seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seq"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ChapterCreate",
        "type": "object"
      },
      "ChapterOut": {
        "properties": {
          "chapter_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Summary"
          },
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "finalized_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalized At"
          },
          "hook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hook"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "done"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "word_count": {
            "default": 0,
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "seq",
          "updated_at",
          "created_at"
        ],
        "title": "ChapterOut",
        "type": "object"
      },
      "ChapterPatch": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "hook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hook"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "ChapterPatch",
        "type": "object"
      },
      "ChapterSaveResult": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "word_count": {
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "word_count",
          "updated_at"
        ],
        "title": "ChapterSaveResult",
        "type": "object"
      },
      "ChapterVersionCreate": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "ChapterVersionCreate",
        "type": "object"
      },
      "ChapterVersionOut": {
        "properties": {
          "chapter_id": {
            "title": "Chapter Id",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "word_count": {
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "chapter_id",
          "word_count",
          "created_at"
        ],
        "title": "ChapterVersionOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI 小说创作工具",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/books/{book}/chapters": {
      "get": {
        "operationId": "list_chapters_api_books__book__chapters_get",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChapterBrief"
                  },
                  "title": "Response List Chapters Api Books  Book  Chapters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Chapters",
        "tags": [
          "chapters"
        ]
      },
      "post": {
        "operationId": "create_chapter_api_books__book__chapters_post",
        "parameters": [
          {
            "in": "path",
            "name": "book",
            "required": true,
            "schema": {
              "title": "Book",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Chapter",
        "tags": [
          "chapters"
        ]
      }
    },
...
    "/api/chapters/{chapter_id}": {
      "delete": {
        "operationId": "delete_chapter_api_chapters__chapter_id__delete",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Chapter Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Chapter",
        "tags": [
          "chapters"
        ]
      },
      "get": {
        "operationId": "get_chapter_api_chapters__chapter_id__get",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Chapter Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chapter",
        "tags": [
          "chapters"
        ]
      },
      "patch": {
        "operationId": "save_chapter_api_chapters__chapter_id__patch",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Chapter Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterSaveResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Chapter",
        "tags": [
          "chapters"
        ]
      }
    },
...
    "/api/chapters/{chapter_id}/versions": {
      "get": {
        "operationId": "list_versions_api_chapters__chapter_id__versions_get",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Chapter Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChapterVersionOut"
                  },
                  "title": "Response List Versions Api Chapters  Chapter Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Versions",
        "tags": [
          "chapters"
        ]
      },
      "post": {
        "operationId": "create_version_api_chapters__chapter_id__versions_post",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Chapter Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChapterVersionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Version",
        "tags": [
          "chapters"
        ]
      }
    },
    "/api/chapters/{chapter_id}/versions/{version_id}/content": {
      "get": {
        "description": "版本正文（「对比当前」功能的数据源）。只读、无副作用。",
        "operationId": "get_version_content_api_chapters__chapter_id__versions__version_id__content_get",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Chapter Id",
              "type": "integer"
            }
          },
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Version Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Version Content Api Chapters  Chapter Id  Versions  Version Id  Content Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Version Content",
        "tags": [
          "chapters"
        ]
      }
    },
    "/api/chapters/{chapter_id}/versions/{version_id}/restore": {
      "post": {
        "operationId": "restore_version_api_chapters__chapter_id__versions__version_id__restore_post",
        "parameters": [
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "chapter_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Chapter Id",
              "type": "integer"
            }
          },
          {
            "description": "资源 id（int64 正整数）",
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "description": "资源 id（int64 正整数）",
              "maximum": 9223372036854775807,
              "minimum": 1,
              "title": "Version Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChapterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Version",
        "tags": [
          "chapters"
        ]
      }
    },
...
  }
}